+++ app/src/main/java/com/thebestory/android/fragment/main/DebugFragment.java
@@ -34,7 +34,7 @@ public class DebugFragment extends Fragment {
      * @return A new instance of fragment DebugFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static DebugFragment newInstance(String param1, String param2) {
+    public static DebugFragment newInstance() {
         return new DebugFragment();
     }