+++ app/src/main/java/com/thebestory/android/fragment/main/stories/NewStoryFragment.java
@@ -4,48 +4,74 @@ import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentTransaction;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.TextView;
 
 import com.thebestory.android.R;
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link RecentFragment.OnFragmentInteractionListener} interface
+ * {@link NewStoryFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link RecentFragment#newInstance} factory method to
+ * Use the {@link NewStoryFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class RecentFragment extends Fragment {
+public class NewStoryFragment extends Fragment {
     private View view;
 
     private OnFragmentInteractionListener mListener;
 
-    public RecentFragment() {
+    private Toolbar toolbar;
+
+    public NewStoryFragment() {
+        // Required empty public constructor
     }
 
     /**
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
-     * @return A new instance of fragment RecentFragment.
+     * @return A new instance of fragment NewStoryFragment.
      */
-    public static RecentFragment newInstance() {
-        return new RecentFragment();
+    public static NewStoryFragment newInstance() {
+        NewStoryFragment fragment = new NewStoryFragment();
+        return fragment;
     }
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        setHasOptionsMenu(true);
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        view = inflater.inflate(R.layout.fragment_main_stories_recent, container, false);
+        view = inflater.inflate(R.layout.fragment_main_stories_new, container, false);
+        toolbar = (Toolbar) view.findViewById(R.id.main_stories_new_toolbar);
+
+        toolbar.setTitle(R.string.main_stories_new_toolbar_title);
+
+        AppCompatActivity activity = (AppCompatActivity) getActivity();
+        activity.setSupportActionBar(toolbar);
+
+        toolbar.setNavigationIcon(R.drawable.ic_close);
+        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                getActivity().onBackPressed();
+            }
+        });
+
         return view;
     }
 
@@ -71,6 +97,31 @@ public class RecentFragment extends Fragment {
     public void onDetach() {
         super.onDetach();
         mListener = null;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+//        FragmentTransaction transaction = getActivity().getSupportFragmentManager().beginTransaction();
+//
+//        switch (item.getItemId()) {
+//            case R.id.main_stories_toolbar_action_new:
+//                transaction.replace(R.id.main_frame_container, newStoryFragment);
+//                break;
+//            case R.id.main_stories_toolbar_action_search:
+//                break;
+//            default:
+//                break;
+//        }
+//
+//        transaction.addToBackStack(null);
+//        transaction.commit();
+
+        return super.onOptionsItemSelected(item);
+    }
+
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+        inflater.inflate(R.menu.main_stories_new, menu);
     }
 
     /**